chalice-cdk/users-service/infra/requirements.txt
aws-cdk.aws-dynamodb
aws-cdk.core
cdk-chalice==0.6.0
